added new options for packing in foamReconstruction
