+++ applications/PUfoam/MoDeNaModels/FoamConstruction/.gitignore
@@ -20,4 +20,7 @@
 *.png
 *.tess
 *.csv
-spherepack
+*.conf
+*.xyzd
+*.nfo
+*.pdf
